📁chore: utils 파일 경로 변경 체험상세에서만 쓰이는 utils 함수를 따로 저징했습니다. Resolve: #205
