+++ src/utils/dateUtils.ts
@@ -72,14 +72,3 @@ export const formatDateForAPI = (date: Date) => ({
   year: date.getFullYear().toString(),
   month: (date.getMonth() + 1).toString().padStart(2, '0'),
 });
-
-// 달 이름 영어로 변경
-export const getMonthNameEnglish = (date: Date): string =>
-  new Intl.DateTimeFormat('en-US', { year: 'numeric', month: 'long' }).format(
-    date,
-  );
-
-// 달 일치여부 확인
-export const isSameMonth = (base: Date, target: Date) =>
-  base.getFullYear() === target.getFullYear() &&
-  base.getMonth() === target.getMonth();
